# webapp/swagger.yaml
basePath: /api/v1
consumes:
    - application/json
definitions:
    APIErrorResponse:
        properties:
            details:
                type: string
                x-go-name: Details
            error:
                type: string
                x-go-name: Error
        type: object
        x-go-package: github.com/cloudbase/garm/apiserver/params
    Address:
        properties:
            address:
                type: string
                x-go-name: Address
            type:
                $ref: '#/definitions/AddressType'
        type: object
        x-go-package: github.com/cloudbase/garm-provider-common/params
    AddressType:
        type: string
        x-go-package: github.com/cloudbase/garm-provider-common/params
    AgentCapabilities:
        properties:
            has_shell:
                type: boolean
                x-go-name: Shell
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    EventLevel:
        type: string
        x-go-package: github.com/cloudbase/garm/params
    EventType:
        type: string
        x-go-package: github.com/cloudbase/garm/params
...
    Instance:
        properties:
            addresses:
                description: |-
                    Addresses is a list of IP addresses the provider reports
                    for this instance.
                items:
                    $ref: '#/definitions/Address'
                type: array
                x-go-name: Addresses
            agent_id:
                description: AgentID is the github runner agent ID.
                format: int64
                type: integer
                x-go-name: AgentID
            capabilities:
                $ref: '#/definitions/AgentCapabilities'
            created_at:
                description: CreatedAt is the timestamp of the creation of this runner.
                format: date-time
                type: string
                x-go-name: CreatedAt
            generation:
                description: |-
                    Generation is the pool generation at the time of creating this instance.
                    This field is to track a divergence between when the instance was created
                    and the settings currently set on a pool. We can then use this field to know
                    if the instance is out of date with the pool, allowing us to remove it if we
                    need to.
                format: uint64
                type: integer
                x-go-name: Generation
            github-runner-group:
                description: |-
                    GithubRunnerGroup is the github runner group to which the runner belongs.
                    The runner group must be created by someone with access to the enterprise.
                type: string
                x-go-name: GitHubRunnerGroup
            heartbeat:
                description: Heartbeat is the last recorded heartbeat from the runner
                format: date-time
                type: string
                x-go-name: Heartbeat
            id:
                description: ID is the database ID of this instance.
                type: string
                x-go-name: ID
            job:
                $ref: '#/definitions/Job'
            name:
                description: |-
                    Name is the name associated with an instance. Depending on
                    the provider, this may or may not be useful in the context of
                    the provider, but we can use it internally to identify the
                    instance.
                type: string
                x-go-name: Name
            os_arch:
                $ref: '#/definitions/OSArch'
            os_name:
                description: 'OSName is the name of the OS. Eg: ubuntu, centos, etc.'
                type: string
                x-go-name: OSName
            os_type:
                $ref: '#/definitions/OSType'
            os_version:
                description: OSVersion is the version of the operating system.
                type: string
                x-go-name: OSVersion
            pool_id:
                description: PoolID is the ID of the garm pool to which a runner belongs.
                type: string
                x-go-name: PoolID
            provider_fault:
                description: |-
                    ProviderFault holds any error messages captured from the IaaS provider that is
                    responsible for managing the lifecycle of the runner.
                format: byte
                type: string
                x-go-name: ProviderFault
            provider_id:
                description: |-
                    PeoviderID is the unique ID the provider associated
                    with the compute instance. We use this to identify the
                    instance in the provider.
                type: string
                x-go-name: ProviderID
            provider_name:
                description: |-
                    ProviderName is the name of the IaaS where the instance was
                    created.
                type: string
                x-go-name: ProviderName
            reserved_for_workflow_job_id:
                description: |-
                    ReservedForWorkflowJobID is the workflow job that claimed this
                    speculative runner. A runner with this set is real work and is never
                    reaped as speculative surplus.
                format: int64
                type: integer
                x-go-name: ReservedForWorkflowJobID
            runner_status:
                $ref: '#/definitions/RunnerStatus'
            scale_set_id:
                description: ScaleSetID is the ID of the scale set to which a runner belongs.
                format: uint64
                type: integer
                x-go-name: ScaleSetID
            speculative:
                description: |-
                    Speculative marks a runner created from a prewarm forecast rather than
                    from a job GitHub had actually queued.
                type: boolean
                x-go-name: Speculative
            speculative_expires_at:
                description: |-
                    SpeculativeExpiresAt is when an unclaimed speculative runner becomes
                    eligible for reaping.
                format: date-time
                type: string
                x-go-name: SpeculativeExpiresAt
            speculative_request_id:
                description: |-
                    SpeculativeRequestID links a speculative runner back to the forecast
                    that created it.
                type: string
                x-go-name: SpeculativeRequestID
            status:
                $ref: '#/definitions/InstanceStatus'
            status_messages:
                description: |-
                    StatusMessages is a list of status messages sent back by the runner as it sets itself
                    up.
                items:
                    $ref: '#/definitions/StatusMessage'
                type: array
                x-go-name: StatusMessages
            updated_at:
                description: UpdatedAt is the timestamp of the last update to this runner.
                format: date-time
                type: string
                x-go-name: UpdatedAt
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    InstanceStatus:
        type: string
        x-go-package: github.com/cloudbase/garm-provider-common/params
    Instances:
        description: used by swagger client generated code
        items:
            $ref: '#/definitions/Instance'
        type: array
        x-go-package: github.com/cloudbase/garm/params
...
    Job:
        properties:
            action:
                description: Action is the specific activity that triggered the event.
                type: string
                x-go-name: Action
            completed_at:
                format: date-time
                type: string
                x-go-name: CompletedAt
            conclusion:
                description: |-
                    Conclusion is the outcome of the job.
                    Possible values: "success", "failure", "neutral", "cancelled", "skipped",
                    "timed_out", "action_required"
                type: string
                x-go-name: Conclusion
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            enterprise_id:
                format: uuid
                type: string
                x-go-name: EnterpriseID
            forge_instance_id:
                format: uuid
                type: string
                x-go-name: ForgeInstanceID
            id:
                description: ID is the ID of the job.
                format: int64
                type: integer
                x-go-name: ID
            labels:
                items:
                    type: string
                type: array
                x-go-name: Labels
            locked_by:
                format: uuid
                type: string
                x-go-name: LockedBy
            name:
                description: Name is the name if the job that was triggered.
                type: string
                x-go-name: Name
            org_id:
                format: uuid
                type: string
                x-go-name: OrgID
            repo_id:
                description: |-
                    The entity that received the hook.

                    Webhooks may be configured on the repo, the org and/or the enterprise.
                    If we only configure a repo to use garm, we'll only ever receive a
                    webhook from the repo. But if we configure the parent org of the repo and
                    the parent enterprise of the org to use garm, a webhook will be sent for each
                    entity type, in response to one workflow event. Thus, we will get 3 webhooks
                    with the same run_id and job id. Record all involved entities in the same job
                    if we have them configured in garm.
                format: uuid
                type: string
                x-go-name: RepoID
            repository_name:
                description: repository in which the job was triggered.
                type: string
                x-go-name: RepositoryName
            repository_owner:
                type: string
                x-go-name: RepositoryOwner
            run_attempt:
                description: RunAttempt is the attempt number of the workflow run this job belongs to.
                format: int64
                type: integer
                x-go-name: RunAttempt
            run_id:
                description: RunID is the ID of the workflow run. A run may have multiple jobs.
                format: int64
                type: integer
                x-go-name: RunID
            runner_group_id:
                format: int64
                type: integer
                x-go-name: RunnerGroupID
            runner_group_name:
                type: string
                x-go-name: RunnerGroupName
            runner_id:
                format: int64
                type: integer
                x-go-name: GithubRunnerID
            runner_name:
                type: string
                x-go-name: RunnerName
            scaleset_job_id:
                description: ScaleSetJobID is the job ID when generated for a scale set.
                type: string
                x-go-name: ScaleSetJobID
            started_at:
                format: date-time
                type: string
                x-go-name: StartedAt
            status:
                description: |-
                    Status is the phase of the lifecycle that the job is currently in.
                    "queued", "in_progress" and "completed".
                type: string
                x-go-name: Status
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
            workflow_job_id:
                format: int64
                type: integer
                x-go-name: WorkflowJobID
            workflow_name:
                description: WorkflowName is the name of the workflow the job belongs to.
                type: string
                x-go-name: WorkflowName
            workflow_run_url:
                type: string
                x-go-name: WorkflowRunURL
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    OSArch:
        type: string
        x-go-package: github.com/cloudbase/garm-provider-common/params
    OSType:
        type: string
        x-go-package: github.com/cloudbase/garm-provider-common/params
...
    RunnerStatus:
        type: string
        x-go-package: github.com/cloudbase/garm/params
...
    StatusMessage:
        properties:
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            event_level:
                $ref: '#/definitions/EventLevel'
            event_type:
                $ref: '#/definitions/EventType'
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
info:
    description: The Garm API generated using go-swagger.
    license:
        name: Apache 2.0
        url: https://www.apache.org/licenses/LICENSE-2.0
    title: Garm API.
    version: 1.0.0
paths:
...
    /enterprises/{enterpriseID}/instances:
        get:
            operationId: ListEnterpriseInstances
            parameters:
                - description: Enterprise ID.
                  in: path
                  name: enterpriseID
                  required: true
                  type: string
            responses:
                "200":
                    description: Instances
                    schema:
                        $ref: '#/definitions/Instances'
                default:
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: List enterprise instances.
            tags:
                - enterprises
                - instances
...
    /instances:
        get:
            operationId: ListInstances
            responses:
                "200":
                    description: Instances
                    schema:
                        $ref: '#/definitions/Instances'
                default:
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: Get all runners' instances.
            tags:
                - instances
    /instances/{instanceName}:
        delete:
            operationId: DeleteInstance
            parameters:
                - description: Runner instance name.
                  in: path
                  name: instanceName
                  required: true
                  type: string
                - description: If true GARM will ignore any provider error when removing the runner and will continue to remove the runner from github and the GARM database.
                  in: query
                  name: forceRemove
                  type: boolean
                - description: If true GARM will ignore unauthorized errors returned by GitHub when removing a runner. This is useful if you want to clean up runners and your credentials have expired.
                  in: query
                  name: bypassGHUnauthorized
                  type: boolean
            responses:
                default:
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: Delete runner instance by name.
            tags:
                - instances
        get:
            operationId: GetInstance
            parameters:
                - description: Runner instance name.
                  in: path
                  name: instanceName
                  required: true
                  type: string
            responses:
                "200":
                    description: Instance
                    schema:
                        $ref: '#/definitions/Instance'
                default:
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: Get runner instance by name.
            tags:
                - instances
...
    /organizations/{orgID}/instances:
        get:
            operationId: ListOrgInstances
            parameters:
                - description: Organization ID.
                  in: path
                  name: orgID
                  required: true
                  type: string
            responses:
                "200":
                    description: Instances
                    schema:
                        $ref: '#/definitions/Instances'
                default:
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: List organization instances.
            tags:
                - organizations
                - instances
...
    /pools/{poolID}/instances:
        get:
            operationId: ListPoolInstances
            parameters:
                - description: Runner pool ID.
                  in: path
                  name: poolID
                  required: true
                  type: string
                - description: List only instances that were created prior to a pool update that changed a setting which influences how instances are created (image, flavor, runner group, etc).
                  in: query
                  name: outdatedOnly
                  type: boolean
            responses:
                "200":
                    description: Instances
                    schema:
                        $ref: '#/definitions/Instances'
                default:
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: List runner instances in a pool.
            tags:
                - instances
...
    /repositories/{repoID}/instances:
        get:
            operationId: ListRepoInstances
            parameters:
                - description: Repository ID.
                  in: path
                  name: repoID
                  required: true
                  type: string
            responses:
                "200":
                    description: Instances
                    schema:
                        $ref: '#/definitions/Instances'
                default:
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: List repository instances.
            tags:
                - repositories
                - instances
...
    /scalesets/{scalesetID}/instances:
        get:
            operationId: ListScaleSetInstances
            parameters:
                - description: Runner scale set ID.
                  in: path
                  name: scalesetID
                  required: true
                  type: string
                - description: List only instances that were created prior to a scaleset update that changed a setting which influences how instances are created (image, flavor, runner group, etc).
                  in: query
                  name: outdatedOnly
                  type: boolean
            responses:
                "200":
                    description: Instances
                    schema:
                        $ref: '#/definitions/Instances'
                default:
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: List runner instances in a scale set.
            tags:
                - instances
...
produces:
    - application/json
security:
    - Bearer: []
securityDefinitions:
    Bearer:
        description: 'The token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"